Suggest `tail -F` instead of `tail -f`
The -F option implies the -f option, but tail will also check to see if the file being followed has been renamed or rotated.  The
file is closed and reopened when tail detects that the filename being read from has a new inode number.  The -F option is ignored
if reading from standard input rather than a file.